requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
dotenv==0.9.9
google-auth==2.47.0
google-genai==1.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
MouseInfo==0.1.3
packaging==25.0
pillow==12.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.2.1
pytweening==1.2.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
